requirements.txt
APScheduler==3.9.1
requests~=2.28.1
pandas~=1.4.4
psycopg2-binary
loguru~=0.6.0